+++ frontend/src/pages/Login.tsx
@@ -13,6 +13,8 @@ import Section2Login from "./sectionLogin/Section2";
 import Section3Login from "./sectionLogin/Section3";
 import LinearIndeterminate from "../Components/LinearIndeterminate";
 import { useNavigate } from "react-router-dom";
+import { AxiosError } from "axios";
+import api from "../services/api";
 
 const Login: React.FC = () => {
   const [value, setValue] = useState(0);
@@ -25,6 +27,48 @@ const Login: React.FC = () => {
 
   const onClick = (value: number) => {
     setValue(value);
+  };
+
+  const handleLogin = (email: string, password: string) => {
+    setLoading(true);
+    api
+      .post("/login", {
+        email,
+        password,
+      })
+      .then((response) => {
+        sessionStorage.setItem("accessToken", response.data.accessToken);
+        const date = Date.now();
+        const expirationTime = date + response.data.expiresIn * 1000;
+        const refreshExpiresIn = date + response.data.refreshExpiresIn * 1000;
+        sessionStorage.setItem("expirationTime", expirationTime.toString());
+        sessionStorage.setItem("refreshToken", response.data.refreshToken);
+        sessionStorage.setItem("refreshExpiresIn", refreshExpiresIn.toString());
+
+        navigate("/timeline");
+      })
+      .catch((error: AxiosError<{ message: string }>) => {
+        if (error.response) {
+          // Resposta com status fora do alcance 2xx
+          const errorMessage =
+            error.response.data?.message ||
+            `Erro de status ${error.response.status}, tente novamente mais tarde!`;
+
+          setAlert(errorMessage);
+        } else if (error.request) {
+          // Requisição foi feita, mas não houve resposta
+          setAlert("Erro de requisição, tente novamente mais tarde!");
+          console.error("Erro na requisição:", error.request);
+        } else {
+          // Algo aconteceu ao configurar a requisição
+          setAlert(`Erro: ${error.message}`);
+          console.error("Erro na configuração:", error.message);
+        }
+        setTimeout(() => {
+          setAlert(null); // Limpa o alerta
+        }, 5000); // 5000 milissegundos = 5 segundos
+      })
+      .finally(() => setLoading(false));
   };
 
   return (
@@ -59,14 +103,14 @@ const Login: React.FC = () => {
             {value === 0 ? (
               <Section1Login onClick={onClick} />
             ) : value === 1 ? (
-              <Section2Login
+              <Section2Login onClick={onClick} handleLogin={handleLogin} />
+            ) : (
+              <Section3Login
                 onClick={onClick}
+                handleLogin={handleLogin}
                 setAlert={setAlert}
                 setLoading={setLoading}
-                navigate={navigate}
               />
-            ) : (
-              <Section3Login onClick={onClick} />
             )}
           </Grid>
           {!isSmallScreen && (
